Added inverse kinematics and simulation
Implemented inverse kinematics into arduino file. Simply set an x,y,z and rotation (-1 to 1 for each axis) and the platform performs the necessary calculations and actuation. The angle on each axis is limited to ±30° because the assembly begins to experience stress beyond this when rigidly mounted; this could potentially be solved by using platform joints with a larger travel range.

A simulation written in processing makes it easy to visualize and check what positions are possible with the platform dimensions. A readout of each actuator's relative length is provided to aid in that. This simulation could be adapted to communicate with the arduino on the fly to set actuator lengths, but currently does not. This sim was adapted from another project located at https://github.com/thiagohersan/memememe/tree/master/Processing/StewartSimulator

+++ src/StewartPlatform/Actuator.h
@@ -40,7 +40,7 @@ class Actuator {
   	
   	void calibrate();
     void calibrate(uint16_t (&settings)[2]);
-  	void setLength(double relativeLength);
+  	void setLength(float relativeLength);
     
   	bool isActuatorReady();
     int getRawPosition();
